Split frc_utils into submodules
